Added a working insertQuestion() method
This is a insert method for Unknown case to insert the question into it
To use: insertQuestion(String LineID, String Question, String
LastFiveQuestions) // please concat the 5 questions into one string
No need to prepare the QID as argument as this is done auto in the
method (see variable stmt2 and currentQid for more)

Using a method for each DB action we want to perform is easier to handle
for us plus setting arguments is easier here.
Sorry George that I deleted code that you created (see comment in
previous coommit).