Integrated UserList into ProjectInterface. It should work like before. + timeout message of previous commit
In ProjectIntreface, userList.getState(String userId) would get back
the state the user is in. userList.setState(String userId, String state)
to change the state of the user.

Consider passing the userList reference to other classes/methods if you
want to modify the state of an User.

(sidenote: it was one hell of a job to separate each user... maybe
ProjectInterface is not as good as just putting everything in KSC, but
the refactoring is too painful now)