Modified ProjectPush to make it allow pushing generic Messages. (also completed javadoc)
In construtor, pass the userId of the user that should receive the
push message.
Then use pushText(String) to push a text message or
pushMessage(Message...) to push up to 5 generic Messages.

A shorthand method is also added to quickly send a text message to a
user without using the constructor. Done by
ProjectPusher.pushText(userId, text);